requirements.txt
streamlit
ultralytics==8.3.156
Pillow
opencv-python-headless
torch==2.7.1
torchvision==0.22.1
gdown
PyYAML
matplotlib